--- docs/contents/2q
@@ -1,9 +1,0 @@
-
-.. small_molecule index
-
-Small molecule
-==============
-
-.. toctree::
-   :maxdepth: 2
-
