--- docs/contents/academia.md
@@ -1,4 +1,0 @@
-https://practicalseries.com/1002-vcs/
-https://carpentries.org/
-https://github.com/carpentries
-
