--- docs/contents/best_practices.md
@@ -1,6 +1,0 @@
-http://education.molssi.org/python-package-best-practices/aio.html
-https://education.molssi.org/python-package-best-practices/
-https://education.molssi.org/
-https://education.molssi.org/python-package-best-practices/index.html
-https://github.com/MolSSI-Education/python-package-best-practices
-
