+++ docs/contents/cards/index.rst
@@ -1,0 +1,9 @@
+
+.. cards index
+
+Sabueso's cards
+===============
+
+.. toctree::
+   :maxdepth: 2
+
