--- docs/contents/data.md
@@ -1,5 +1,0 @@
-
-
-https://setuptools.readthedocs.io/en/latest/pkg_resources.html#resourcemanager-api
-https://setuptools.readthedocs.io/en/latest/userguide/datafiles.html
-https://setuptools.readthedocs.io/en/latest/pkg_resources.html#resourcemanager-api
