--- docs/contents/devtools.md
@@ -1,8 +1,0 @@
-# devtools
-
-## scripts
-
-```bash
-python create_conda_env.py -n Test-UIBCDF-Library -p 3.7 ../conda-envs/development_env.yaml
-```
-
