--- docs/contents/documentation/index.rst
@@ -1,30 +1,0 @@
-.. documentation master file
-
-Documentation
-=============
-
-XXXXX
-
-.. _libraries-supported:
-
-.. toctree::
-   :name: documentation
-   :caption: Documentation
-   :maxdepth: 1
-
-   sphinx.md
-   conf.md
-
-.. toctree::
-   :caption: API Documentation
-   :maxdepth: 2
-
-   api
-
-Glossary, indices and tables
-============================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
-
