--- docs/contents/installation_and_quick_guide/index.rst
@@ -1,12 +1,0 @@
-
-.. intro intro
-
-Installation and Quick Guide
-============================
-
-.. toctree::
-   :maxdepth: 1
-
-   Installation.md
-   Quickstart.ipynb
-
