+++ docs/contents/protein/index.rst
@@ -7,3 +7,7 @@ Protein
 .. toctree::
    :maxdepth: 2
 
+   is_protein.ipynb
+   get_function.ipynb
+   forms.ipynb
+
