+++ docs/contents/showcase/index.rst
@@ -1,0 +1,8 @@
+.. _showcase:
+
+Showcase
+========
+
+.. toctree::
+   :maxdepth: 1
+
