--- docs/contents/tags_and_releases.md
@@ -1,5 +1,0 @@
-https://docs.github.com/en/github/administering-a-repository/releasing-projects-on-github/about-releases
-https://docs.github.com/en/github/administering-a-repository/releasing-projects-on-github/managing-releases-in-a-repository
-https://dev.to/neshaz/a-tutorial-for-tagging-releases-in-git-147e
-https://git-scm.com/book/en/v2/Git-Basics-Tagging
-
