+++ docs/contents/tools/index.rst
@@ -5,12 +5,16 @@ Tools
 =====
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 1
 
+   string_entity_name/index.rst
+   string_entity_query/index.rst
    string_pdb_id/index.rst
+   string_fasta_text/index.rst
    string_uniprot_id/index.rst
    database_RCSB_PDB/index.rst
    database_UniProtKB/index.rst
    file_pdb/index.rst
+   sabueso_ProteinDict/index.rst
    sabueso_UniProtKB_XMLDict/index.rst
