+++ docs/contents/tools/string_entity_name/index.rst
@@ -1,0 +1,13 @@
+
+.. tools string_entity_name index
+
+string entity name
+==================
+
+.. toctree::
+   :maxdepth: 2
+
+   string_is_entity_name.ipynb
+   is_protein.ipynb
+   to_string_uniprot_id.ipynb
+
