+++ docs/contents/tools/string_entity_query/index.rst
@@ -1,0 +1,13 @@
+
+.. tools string_entity_query index
+
+string entity query
+==================
+
+.. toctree::
+   :maxdepth: 2
+
+   string_is_entity_query.ipynb
+   is_protein.ipynb
+   to_string_uniprot_id.ipynb
+
