+++ docs/contents/tools/string_fasta_text/index.rst
@@ -1,0 +1,11 @@
+
+.. tools string_fasta_text index
+
+string fasta text
+==================
+
+.. toctree::
+   :maxdepth: 2
+
+   get_sequences.ipynb
+
