+++ docs/index.rst
@@ -6,21 +6,30 @@ Sabueso
 .. image:: https://anaconda.org/uibcdf/molsysmt/badges/license.svg
    :target: https://github.com/uibcdf/MolSysMT/blob/master/License
 
-Sabueso ("Hound" in Spanish) hunts any kind of information on the web related with your molecular
-system. You can also ask Sabueso to look for a specific molecule and Sabueso will give you all data found in the online databases as a Python object ("card"). Or you can take advantage of Sabueso's hunting tools to build your own workflow.
+Sabueso ("hound" in Spanish) hunts any kind of information on the web related with your molecular
+system. You can ask Sabueso to look for a specific molecule or system, and Sabueso will give you back useful data from online databases as a Python object ("card"). Or, instead, you can take advantage of Sabueso's hunting tools to build your own workflow.
 
-Here you can find just a short list with some of kind of things you can do with Sabueso:
-
-- Bla    
-- Bla    
-- Bla    
-
-Have a look to the "Quick guide" section and give it a try!
+Have a look to the :ref:`showcase section<showcase>` and give it a try!
 
 .. warning:: |IntroWarningText|
 
 .. |IntroWarningText| replace::
         At this moment Sabueso is a work in progress project maintened by the UIBCDF Lab. If you think this library is useful to you, use it with caution. Any feedback is welcome.
+
+.. toctree::
+   :name: about
+   :caption: About
+   :maxdepth: 2
+
+   contents/about/what.md
+
+.. toctree::
+   :name: installation_and_showcase
+   :caption: Installation and showcase
+   :maxdepth: 2
+
+   contents/installation/Installation.md
+   contents/showcase/index.rst
 
 .. toctree::
    :name: user_guide
@@ -29,6 +38,7 @@ Have a look to the "Quick guide" section and give it a try!
 
    contents/installation_and_quick_guide/index.rst
    contents/intro/index.rst
+   contents/cards/index.rst
    contents/ion/index.rst
    contents/water/index.rst
    contents/cosolute/index.rst
@@ -38,6 +48,7 @@ Have a look to the "Quick guide" section and give it a try!
    contents/protein/index.rst
    contents/rna/index.rst
    contents/dna/index.rst
+   contents/native/index.rst
    contents/tools/index.rst
 
 .. toctree::
