+++ sabueso/tools/string_entity_name/__init__.py
@@ -1,2 +1,3 @@
 from .is_string_entity_name import is_string_entity_name
+from .to_string_uniprot_id import to_string_uniprot_id
