+++ sabueso/tools/string_uniprot_id/to_sabueso_UniProtKB_XMLDict.py
@@ -1,8 +1,8 @@
-import xmltodict
-import requests
-from sabueso.tools.database_UniProtKB import is_accessible as UniProtKB_is_accessible
-
 def to_sabueso_UniProtKB_XMLDict(item, indices='all'):
+
+    import xmltodict
+    import requests
+    from sabueso.tools.database_UniProtKB import is_accessible as UniProtKB_is_accessible
 
     if item.startswith('uniprot_id:'):
         item=item[11:]
